Render "incoming" commit in conflict area
Summary:
If you hit a conflict while rebasing A onto B, previously we only showed the conflicts on top of B (the "local" side of the merge). But it's hard to tell what the "incoming" side of the merge is.

Now that we know the commit hashes of the merge sides, we can render the commit in a preview style with an arrow to indicate the merge. We can also add labels to show the merge side names.

I added a <CommitLabel> mechanism to show styled labels above commits. Right now it's only used for this, but this could be coopted for other uses too.

Note: since I do this commit label by hash, it will show "incoming" next to the commit in the tree normally, too. I think that's a kind of nice feature, actually.

In a follow up diff, I plan on making "local" / "incoming" names configurable once and for all so you don't get tripped up on the naming. I recognize that it's already inconsistent with "Deleted in destination".

Reviewed By: zzl0

Differential Revision: D56861540

fbshipit-source-id: cb3611aeeac7fcf055ba1f7224099693e849b901