gitimport: Make catch-up part of full repo import > 100x faster
Summary:
A common use for the `gitimport` tool is to catch-up import for a repo that was previously imported.

Before this diff, this use-case was poorly supported for large repos as the part of the tool that
checks which commits are already imported was inefficient.

For instance, for `chromium/src` which has ~2.3M commits and is already imported, reaching the point
where we upload bookmarks used to take over 2 hours.

After this change, we get to the same point in ~1 min 30:
the first 45s to list the commits from git which we already used to do, then another 45s to check which
mappings already exist, which used to take over 2 hours.

## Details

Instead of checking each mapping 1 by 1, do batched queries of 100k commits at a time in the `bonsai_git_mapping`.

Do these queries in parallel.

Instead of calling the git client twice, once to enumerate the commits and once to count them (both operations taking about 45s), call it only once and store the commits in memory.
Re-use this for the count and for the `bonsai_git_mapping` lookups.

Reduce the number of times the lock on `acc` has to be acquired and released by preloading all mappings in a functional way, then
extending acc in one go with all that we found. (This account for ~8mins -> 45s speed up for that section).

Stop logging each commit that already exists to reduce log-spew. Instead, only log the aggregate of how many already exist.

Reviewed By: markbt

Differential Revision: D56854649

fbshipit-source-id: 8bd3f055ed07726bef3c675344a13ec7a7168ef9