Update labelling sides of a conflict
Summary:
People have always been confused with the terminology used for the "sides" of a conflict:
- local vs incoming
- local vs other
- dest vs source
- theirs vs yours
- rebasing onto vs being rebased

~~If we make this configurable, then we can have people customize the names in the way they'd like.~~

~~This diff starts by using a local storage config to replace all these names in the UI.~~

~~In a separate diff, we'll make this state configurable.~~

~~The default is set to source and dest, but we may want to change this. Although I did find that having a very long name by default can make it a bit hard to parse.~~

~~There is one problem with this approach: this won't affect the conflict markers shown inside the file when you start resolving.
But I think this is controlled by an sl config (haven't really checked...), so we could set that config for our `rebase` operations and use whatever value you've set in the UI, so after you set it it would be used for any rebases triggered by ISL. That seems like a good way for it to work. Then it should still only impact ISL, while still listening to your preference.~~

Update:
Just consolidate where we define the labels so we can easily change them later (or make them configurable if we later decide it's the right thing to do).

update the labels to be:
```
"source - being rebased"
```
and
```
"dest - rebasing onto"
```

Reviewed By: quark-zju, zzl0

Differential Revision: D56862394

fbshipit-source-id: 7c49a3b64a4e8051eecb46b59239463c63d3c14a