populate fetchedSource to trace hg
Summary:
## Background
`SaplingBackingStore` is now sending `request` to Sapling with `RemoteOnly` or `LocalOnly`. Therefore `SaplingBackingStore` knows the source of fetching each request. We need to save this info in `ObjectFetchContext` then we can send them to `HgImportTraceEvent::finish` and `traceBus_` can publish this info in `trace hg` command.

## This diff
Now that context have the value of `FetchedSource`, we can populate it in the `HgImportTraceEvent::finish` event. It is obvious that `HgImportTraceEvent::start` and `HgImportTraceEvent::queue` events doesn't know the `FetchedSource` because it is not fetched yet.
Also, this diff modified `trace_stream.cpp` to show the FetchedSource by emoji in `trace hg` command when it runs with `--verbose` argument
```
static const auto kLocalFetchedEmoji =
    reinterpret_cast<const char*>(u8"\U0001F4BB"); // 💻
static const auto kRemoteFetchedEmoji =
    reinterpret_cast<const char*>(u8"\U0001F310"); // 🌐
static const auto kUnknownFetchedEmoji =
    reinterpret_cast<const char*>(u8"\U0001F937"); // 🤷
```

Differential Revision: D56283112

fbshipit-source-id: 8b7d461ff283b6aba8d894f9ec809e6a974c2201