set FetchedSource in ObjectFetchContext
Summary:
## Background
`SaplingBackingStore` is now sending `request` to Sapling with `RemoteOnly` or `LocalOnly`. Therefore `SaplingBackingStore` knows the source of fetching each request. We need to save this info in `ObjectFetchContext` then we can send them to `HgImportTraceEvent::finish` and `traceBus_` can publish this info in `trace hg` command.

## Thid diff
Adding a new field to `ObjectFetchContext` to keep the source of fetch for each request. The source can be LOCAL, REMOTE, or UNKNOWN

Reviewed By: jdelliot

Differential Revision: D56332989

fbshipit-source-id: d684593c26f8b97276130bf3972479e6b18dabaf