refactor SaplingImportRequest to keep ObjectFetchContext
Summary:
## Background
`SaplingBackingStore` is now sending `request` to Sapling with `RemoteOnly` or `LocalOnly`. Therefore `SaplingBackingStore` knows the source of fetching each request. We need to save this info in `ObjectFetchContext` then we can send them to `HgImportTraceEvent::finish` and `traceBus_` can publish this info in `trace hg` command.

## This diff
`ObjectFetchContext` is not attached to `request` and when the `request` is fulfilled we cannot update the context. This diff adds a field `context_` to the `SaplingImportRequest` class then each request can access its own context.

Reviewed By: genevievehelsel

Differential Revision: D56221059

fbshipit-source-id: 2c4f73bfa86912c3189204a3c8c01a528dab0715