Fix fetchTree duration ODS in edenStats
Summary:
`SaplingBackingStoreStats::fetchTree` is the duration that is supposed to record the fetching time (in Microsecond) when a tree/root tree is fetched from BackingStore. The `SaplingBackingStore.cpp` went through multiple merging and refactoring. All those changes cause this duration to not work accurately. To find that this duration is not accurate first we have to look at how `getTree`, `getRootTree`, and `importManifestForRoot` works in `SaplingBackingStore.cpp`

## Fetching Tree flow
We have three functions in `SaplingBackingStore.cpp` that fetch Tree from BackingStore:
1- `SaplingBackingStore::getTree` : It first checks local, if it is not found prepare a request and enqueue it.  Requests will be dequeued with `processTreeImportRequests` and send to BackingStore with `getTreeBatch` and for the failed one we call `retryGetTree`  which it calls `retryGetTreeImpl`
   processTreeImportRequests
    ↳ getTreeBatch
    ↳ retryGetTree
      ↳ retryGetTreeImpl
2- `getRootTree`: Based on the commitID StoreResult, it calls `importTreeManifest` or `importTreeManifestImpl`. Finaly in the `importTreeManifestImpl` the tree fetched from BackingStore, and if it is not found it will be retried by `retryGetTree`  and then `retryGetTreeImpl`
  getRootTree
   ↳importTreeManifest
   ↳importTreeManifestImpl
   ↳ retryGetTree
     ↳ retryGetTreeImpl
3- `importManifestForRoot`: It calls `importTreeManifestImpl` and if it is not found in the first try it goes to same direction for  `retryGetTree`  and then `retryGetTreeImpl`
   importManifestForRoot
    ↳importTreeManifestImpl
    ↳ retryGetTree
      ↳ retryGetTreeImpl

## Current fetch_tree duration records
Currently, we record fetch_tree duration in `processTreeImportRequests`,
`importTreeManifestImpl`, and `retryGetTree`. This is not accurate. Because `processTreeImportRequests` already record the time in `retryGetTree`, and recording this duration again in `retryGetTree` is a duplicated data.

## This diff solution
To not duplicating fetch_tree duration records, we should remove records from `importTreeManifestImpl` and `retryGetTree`. On the other hand, to confirm that the tree fetches which are running through `getRootTree` or  `importManifestForRoot` are recording correctly, this diff brings the records into these two functions.

This doc explains all the SaplingBackingStore ODS counters/duration after this stack changes: https://docs.google.com/document/d/1o355e8JGvq3fBYFD738Jkx9tPSWzH4bmxcM8YRKMziU/edit?usp=sharing

Reviewed By: jdelliot

Differential Revision: D56530480

fbshipit-source-id: 486e38c7d0fe4e9d1be77d2cf15f2752bad29657