avoid calling getStats() in some ObjectCache fb303 callbacks
Summary: The current state of this code causes an infinite loop if trying to use `getStats` for the `eden stats` cli in D56453268. This diff introduces a specific helper function to extract these singular values. Also, we don't need to collect the whole of the `Stats` objects anyways to feed these two dynamic counters anyways.

Reviewed By: kmancini

Differential Revision: D56728227

fbshipit-source-id: 22523f97a6864464beed875e82fef1d99348fe2a