Implement suffix querying
Summary:
This Diff implements the EdenAPI Suffix Query handler.

The handler extracts the changeset from the eden context and uses it to call find_files_with_bssmv3.

The output of that function is a stream of MPaths. These are transformed into SuffixQueryResponses and returned in and async stream.

This Diff also implements several traits for CommitID to help in the functionality of above.

Reviewed By: markbt

Differential Revision: D56433194

fbshipit-source-id: 28085e6c345cb33bfa33f9cef9c9ed5708d483dd