Store submodule dependencies in LargeToSmall direction
Summary:
We'll be needing it soon to support backsyncing and I'm working on fixing a bug that's partly caused by the fact that `CommitSyncRepos::new` takes submodule deps but drops it silently when creating `LargeToSmall`.

The call to `get_submodule_deps` then always returns `NotNeeded`.

This is the case because when I initially implemented this, we were under the assumption that we would only be supporting forward syncing.

Reviewed By: mitrandir77

Differential Revision: D56881703

fbshipit-source-id: bb90fbc39b25d4fadc5342dafa35c02f93b25bd0