revsets: search Git diffs locally if phrevset.graphqlonly is set to False
Summary:
Even with the changes from D56443510 it's not enough to make Sapling look for
local Git commits in case there is a newer Phabricator version. Sapling commits
do not have this issue since they are backed up by commitcloud, so those usually
get autopulled if Phabricator's version is more recent than the local one.

This diffs changes that for Git repos, showing a warning telling users to use `jf get` to get the latest version since that error message should be more useful for them.

Reviewed By: muirdm

Differential Revision: D56447240

fbshipit-source-id: 54025ac7f330c629b3b804a3b9dc1a3013e29d83