revsets: make Phabricator revsets work for the most part again for Git repos
Summary:
Some time ago Phabricator revsets were usable on Git repos, but seems like
recently something broke and it's not possible to use them for finding
local commits anymore, which this diff fixes. Additionally, this diff
makes it possible again to find public commits by Phabricator revsets,
which seems to have been broken for a much longer time.

There are two root causes for these issues:
- `remotefilelog.reponame` has the name of Git repos as `reponame` instead
of `reponame.git`, which in turn causes `sl` to error out with issues like
`abort: D123 is for repo foo/bar/baz.git', not this repo ('baz')`
(fixed in D56444313)
- The logic for resolving Git revisions from Phabricator was written 10 years
ago, and depends on using sshpeer which is not really supported anymore: D1458491

Reviewed By: muirdm

Differential Revision: D56443510

fbshipit-source-id: 14e176cac5de1216b1e170d0863d7b7c6382df59