avoid crashing when invalid FilteredObjectIDs are parsed
Summary:
# Context

Genna discovered that passing junk to `eden debug blob` on a FilteredFS repo would cause the Eden daemon to crash. This is because `XCHECK_EQ()` eventually resolves to `XLOG(FATAL, ...)` if the equality check fails[1]. We don't actually want to crash in this case, since a user can pass non-sense to 'eden debug blob' whenever they'd like

# The Fix

We should just throw an invalid argument exception if the ID we get passed in looks like garbage. This causes the thrift call to fail instead of making the daemon crash.

In addition, I moved the `XCHECK()` lower in the function to be an `XDCHECK()` so that we only crash the daemon in debug mode (not production). This is fine, since that check is meant to be hit during development, not in produciton!

Reviewed By: genevievehelsel

Differential Revision: D56846354

fbshipit-source-id: 3623eb05ccff83b4186b7f4e4363fa769df872bd