add the option for log to use the committer date instead of the usual date
Summary:
In repositories imported from Git there are two dates: the author date and committer date.  Depending on whether they were imported into Mononoke first (which uses the same style as hg-git), or imported into Sapling directly, these dates are mapped differently.

In the cases where Mononoke did the import, we may want to configure repositories to prefer the committer date, rather than the author date, which is what is used by default.  Add a new config option `log.use-committer-date` which does exactly this.

We also extend the JSON format for `hg log` to include both author date and committer date if they are available.  This means tooling that knows specifically which one it wants can be modified to fetch exactly that.

Reviewed By: quark-zju

Differential Revision: D56941671

fbshipit-source-id: 0d68a333caf06b8b82e08b581f06df6394520992