Fix for git pull with tags in Mononoke GRit server
Summary:
As seen in the previous diff, when the request from the client includes tag Object Ids in the `WANTS` section, we just consider the commit pointed to by the tag and completely ignore the tag itself. As a result, the output tag stream sent as part of the server response does not include the requested tags. The changes in this diff ensure that every tag that matches either of the following constraint gets sent in the response to the client:
- If a tag points to any commit that is part of the response, then that tag gets sent
- If a tag is explicitly requested by the client through the `WANTS` section, then that tag gets sent

Reviewed By: mitrandir77

Differential Revision: D56980436

fbshipit-source-id: 5d92dff784e212ec71b6b78c9f50e0877731c8c1