Fix for delta cycles in partial pull / fetch through Mononoke GRit server
Summary: When pulling the latest changes for `whatsapp/server` through `Mononoke GRit` server, we encountered an error which was essentially a result of a cycle included in the packfile. Git has weird set of conditions in which it allows cycles and certain other conditions in which it doesn't allow cycles. To avoid bumping into this problem in the future again, let's build in cycle-detection logic which prevents from sending objects like A and B where A is expressed as delta of B and B is expressed as delta of A.

Reviewed By: mitrandir77

Differential Revision: D56981007

fbshipit-source-id: dad6e11460531176b805d94c06f857377a850dcd