Add "Clear commit message" button in fill commit message context menu
Summary:
Next to the "fill commit message" section in the commit-mode commit info view, add a dropdown button that lets you choose to clear the commit message. This is useful if you use the fill commit message button but want to go back and not have all of the fields filled in.

in the future, I suspect we'll add more fill methods like filling from a task/diff, and if so we may need to move some fill methods to this context menu as well for space savings.

Note: we only show this for "commit" mode because amend mode alredy has a "cancel" button to go back to the original message. But "cancel" doesn't really work for commit mode.

Reviewed By: zzl0

Differential Revision: D57013039

fbshipit-source-id: 201c6e380defd2f53fc67b6ef0309bcba829e163