tests: add a target for running run-tests.py with EdenFS
Summary:
It's sometimes annoying to run EdenFS .t tests without having autofix for them, which is why this diff was created.

In order for EdenFS .t tests to run properly, some env vars have to be set. Ideally we would modify the existing `run_tests_py_eden` target, but the problem with this is that Buck doesn't allow Python binary targets to take env vars as input.

Reviewed By: zzl0

Differential Revision: D57010488

fbshipit-source-id: 9155f39c2ba479bf763a24075316b439453ab6e1