Use realpath for displayed cwd, compute on server
Summary: Previously, we sent the repo root and the relative cwd to have the UI combine into a display cwd. We can instead use realpath and join them on the server to get a label for the UI to use. We don't even really need to send the repoRelativeCwd, just the label.

Reviewed By: quark-zju

Differential Revision: D57116811

fbshipit-source-id: 7f1f9d6f8be3751daf5242a7fccbde9c8e6008f8