Show cwd button label with full path relative to repo root
Summary:
Instead of showing the basename of the cwd in the cwd dropdown, show the repo root basename + the relative path to the cwd.

For example, "sapling/addons/isl" instead of just "sapling" (or just "isl").

This makes it more clear what your cwd is within the repo, which can give you more information in some cases.

Reviewed By: quark-zju

Differential Revision: D57083698

fbshipit-source-id: 3832eb7cf8504f2545bff2e9745de78940086e6d