Ensure platform-specific server message handling always has a RepositoryContext
Summary:
If a platform has its own dedicated message handling, we previously might pass in a null RepositoryContext. However, we are always be able to create a RepositoryContext from the information we have.

Constructing one will make it easier to use platform-specific messages, for example to determine more info about the cwd.

Reviewed By: quark-zju

Differential Revision: D57083697

fbshipit-source-id: 6d3f3cfcca1ae76d7075156e462187891fceaa08