Use repo-relative paths for cwd selector
Summary:
The previous cwd selector used minimalDisambiguousPaths for the cwd selector. This meant that typically you'd only see the basename of the cwd as an option for the cwd.

Instead, we can first find the repo root for each cwd option, and thus render the cwd as relative to that rpeo root. This also lets us know ahead of time which cwd options are invalid, so we can disable them.

I also hide invalid options from the quick cwd picker.

We could later augment this data too, for example labelling git or sl repos.

Reviewed By: quark-zju

Differential Revision: D57083696

fbshipit-source-id: 5e7a0e0f1e96019e9512d711e773d97e2d62c352