Add checkbox next to sections of sectioned file tree
Summary:
Add a checkbox next to sections so you can toggle all files within a section at once. This is much more convenient if you need to make a partial commit from files in just one section.

In the next diff, we'll also indent these nested files so the checkboxes are differentiated better

Reviewed By: quark-zju

Differential Revision: D57121564

fbshipit-source-id: 19b96143042f35675318b6cc232959fd7648f264