Add file selection checkboxes to levels of file tree
Summary:
If you view uncommitted changes in "tree" mode, we should have checkboxes on each level to let you select/deselect all files "under" that level.

We can use the "indeterminate" state if there's a mix of selected files in a level of the tree, or if some files are themselves partially selected (via chunk selection)

This involves iterating all files inside a level in the tree, both to calculate the checked state and to select/deselect on click.

I added a few iterator utils so we don't need to make new arrays out of all the file paths when computing this (since we would do that for each level, that could be somewhat expensive). [These helpers are actually coming soon to js itself, they're at stage 3!](https://github.com/tc39/proposal-iterator-helpers?tab=readme-ov-file) Can't depend on that just yet though.

Reviewed By: muirdm

Differential Revision: D57124747

fbshipit-source-id: 3ba0bf5d593b153a2d6beef69a9592944e6e9e3f