Log to both scuba and terminal in cross_repo_sync and forward syncer
Summary:
Introduce helpers that will log to both slog (i.e. terminal) and Scuba.

This will be useful when to debug/follow the initial import command. It might also help debug cross_repo_sync stuff in general.

This diff is supposed to be a no-op! I'm updating/adding logs in the next diff.

Reviewed By: andreacampi

Differential Revision: D56925258

fbshipit-source-id: b914c59964ca9be8b9068336713d0953042c1bd1