CursorSerialization 1/n: Introduce CursorSerializationWrapper and adapter
Summary:
Thrift Cursor Serialization is a new serde API for C++ that permits reading and writing Thrift structs encoded with the Thrift Binary protocol by directly manipulating the serialized buffer rather than writing into a generated struct as a first step.
This diff adds a typed wrapper that manages the serialized buffer and an adapter for using this wrapper in the signature of service methods.

Reviewed By: thedavekwon

Differential Revision: D56495330

fbshipit-source-id: d744ecb045bb0557c9420a239ef4caea0c668ae1