CursorSerialization 1.5/n: Add validator for use of adapter
Summary:
Thrift's binary and compact protocols are designed for complete sequential deserialization. Fields in a struct and elements in a container are simply concatenated with no delimiter as the expectation is that each wire type knows how to determine its own length (which is almost always variable). In the process of deserializing one field, the cursor is advanced to the start of the next field. However, computing the distance the cursor will advance is often as expensive as deserializing the wire type, and is required in order to skip a field.
This poses a problem when the cursor serialization adapter is applied to a field in a struct (including the RPC request struct) or an element of a container, as the adapter needs to consume the right number of bytes so the deserialization of the next field will work correctly. Naively calling skip to perform this computation would eliminate much of the CPU savings.
There are some contexts where advancing the cursor correctly doesn't matter, namely when the adapted struct makes up the entire body of an RPC request or response or a serialized buffer. Restricting the use of this adapter to those contexts allows the adapter to bypass the skip computation. This validation enforces that restriction.

Reviewed By: thedavekwon, vitaut

Differential Revision: D56835476

fbshipit-source-id: 0aea25315d73b02514179f0946ae424260f80256