tests: make `(eden !)` statements on .t tests work properly
Summary:
Debugruntest does `hghave.py` checks at a couple of points:
- When converting .t tests into Python code with `transform()`, which affects `#if` code blocks as well as `#require`s
- When using `(myfeature !)` line matching

The former is run *before* actually executing anything in the test (hence the "transform" part), and thus the actual environment variables that `debugruntest` receives are the ones that are used for checking features.

The later is run *after* commands are run, and output from commands is being compared. So at this point, the env vars being used are not necessarily the ones that are checked in the first step. In most checks this is not an issue, but we have a few like `fsmonitor` (which checks for Watchman use):

https://www.internalfb.com/code/fbsource/[272c9d8a9228738733d213384257f83d13fcdc59]/fbcode/eden/scm/sapling/testing/t/hghave.py?lines=524-526

The way this is solved on existing tests is to simply append that once again on the hg debugruntest extension:

https://www.internalfb.com/code/fbsource/[272c9d8a9228738733d213384257f83d13fcdc59]/fbcode/eden/scm/sapling/testing/ext/hg.py?lines=139

Rewriting debugruntest so that the checks for those two steps would happen at the same time would probably be a better idea, but this would probably require rewriting some chunks of debugruntest. Instead of doing that, let's just do for EdenFS the same we have done for Watchman in the past.

Also, the reason `(eden !)` statements were working for some of our existing EdenFS tests is that hghave checks get cached, and since `#require` and `#if` are used for checking for EdenFS compatibility in some of them, once passing using the env vars from the first type mentioned above the checks from the second type weren't necessary.

Reviewed By: muirdm

Differential Revision: D57180999

fbshipit-source-id: 3eca355920111b26641e6fd306453ed3bb4e420a