gotham_ext: use a common base scuba sample builder
Summary:
Due to conflicting abstractions, the HTTP services in Mononoke have *two* scuba sample builders:

* The regular scuba sample builder in `CoreContext`.  This is used by most Mononoke code to log events, usually by calling `ctx.scuba().clone()`.  It is stored in the `RequestContext` in the Gotham `State`.

* The `ScubaMiddleware` scuba sample builder.  This is constructed during inbound request processing and used to log request completed and cancelled messages.  It is stored in `ScubaMiddlewareState` in the Gotham `State`.

Currently, only the second of these is populated with request information, like the client entry point or identities, or request path.  This means code using the first method will not log these fields.

Ideally there would be only one such builder, but that is a larger refactor.  For now, mitigate things by re-using the scuba sample builder from the `CoreContext` when constructing the scuba middleware, and replacing the scuba sample builder on the `CoreContext` with the populated one.

Reviewed By: zzl0

Differential Revision: D57210802

fbshipit-source-id: d3c2fa2e78cef8da0dea129dc751970724ecea11