Change return type of top-level submodule expansion function
Summary:
## This stack
See D57204338.

## This diff
Currently we only call `expand_and_validate_all_git_submodule_file_changes` if we expect to get a bonsai.
But in later diffs we'll call this function if the small repo has action set to EXPAND, regardless of direction.

If the direction is backsyncing, we'll return the result of `rewrite_commit`, which is an optional Bonsai. This is necessary for the callers to do the right thing.

This diff does the following:

- Changes the return type of `expand_and_validate_all_git_submodule_file_changes`
- Renames it to `rewrite_commit_with_submodule_expansion`, since it will basically do that, but handle submodule expansion and validation if needed.
- Rename small_repo to source_repo, because it won't always be called with small_repo

Reviewed By: mitrandir77

Differential Revision: D57205271

fbshipit-source-id: 28e9aced72b2337893cbd5c80c13cedd34bc57df