Add option to specify the number of lines of unified context
Summary:
This diff adds an optional argument to specify the number of lines of unified context to show around differences. Specifically, this controls the option `context` of `CommitFileDiffsParams`:

https://www.internalfb.com/code/fbsource/[a68db4bdf6284544725693162650ae766fb783b3]/fbcode/eden/mononoke/scs/if/source_control.thrift?lines=1180%2C1190-1191

This will allow me to debug the response of the Mononoke service with different lines of context.

Other implementations of a diff command typically expose this option as `-U` or `--unified`. I wasn't sure whether to follow this convention or name the argument `--context` to mimic the Thrift struct.

```
$ hg diff --help
 -U --unified NUM          number of lines of context to show

$ diff --help
 -u, -U NUM, --unified[=NUM]   output NUM (default 3) lines of unified contex

$ git diff --help
 -U --unified NUM          number of lines of context to show
```

Reviewed By: mitrandir77

Differential Revision: D57217689

fbshipit-source-id: 19d07e0429113c76668cfd5e308e048678f249b5