Provide support for deltas in filtered clones
Summary: When serving a repo in filter mode, some subset of objects are excluded by the server while sending the response over to the clients. If any of these excluded objects are used as bases for deltas, we end up creating a packfile with incomplete deltas. This diff ensures that we only use deltas when both the target and the base object pass the necessary filters specified by the client.

Reviewed By: mitrandir77

Differential Revision: D57246404

fbshipit-source-id: b7dadbef78caee471f34e754643a1327dacc5668