status: don't rely on ignore matcher checking manifest
Summary:
I'm trying to make the ignore_matcher not be manifest aware. In general the treestate is a cache on top of the manifest, so we can use the treestate to determine if a file is tracked without needing to consult the manifest. Tweak a couple spots to do that. In particular, don't assume the `matcher` has already filtered out ignored files, and don't assume the `ignore_matcher` knows that a tracked file doesn't count as ignored.

I will remove the manifest matcher in later diff.

Reviewed By: quark-zju

Differential Revision: D57171617

fbshipit-source-id: cd25497a51ac3ddc88bd59e2ae3390ecc990b4cd