status: stop using ManifestMatcher
Summary:
The ManifestMatcher is a Matcher which matches a file if the file exists in the manifest. Makes sense conceptually, but in practice it leads to serial tree fetches since the Matcher is interrogated one file at a time.

In general we shouldn't need the ManifestMatcher since the dirstate can tell us whether a file is tracked or not. In the case of eden, eden tells us whether the file is tracked or not. In previous diffs I tweaked the FileSystem implementations to not assume the `matcher` and `ignore_matcher` are manifest aware, and in this diff I remove the manifest matcher from the `matcher` and `ignore_matcher`.

I also made the `matcher` no longer ignore matcher aware. It doesn't seem like that was required, and now things are easy to reason about since there is less "overlap" between the `matcher` and the `ignore_matcher`. The `matcher` is the user specified patterns plus any sparse matcher. The `ignore_matcher` is the GitIgnorematcher plus anything outside the sparse matcher.

Hopefully this will avoid edge cases that trigger a lot of serial tree fetches.

Reviewed By: sggutier

Differential Revision: D57171618

fbshipit-source-id: b208b29bbd63e42975716d003f68b97841c6d1c1