Automatically run custom merge tool
Summary:
Previously, we added support for custom configured merge tools. But we didn't run them by default, instead you had to manually click a button to open the tool.

Instead, if we don't set `SL_MERGE=:merge3`, sl will automatically use the configured merge tool. That's probably a better behavior.

Using the merge tool directly during the rebase (as opposed to a separate button press) also has the advantage of not hitting weirdness if you resolve some files before doing the `resolve --all`, which can reset your resolutions.

The only reason we overwrote SL_MERGE was to use `:merge3` instead of `:merge` by default, so you get 3-way merge tools when used in VS Code. We still kind of want this behavior by default I think, so we can check if `ui.merge` is set to `:merge` (the default) and use that to determine if we should set `:merge3` instead, and otherwise don't set these configs.

We definitely want to avoid the CLI using vim or other CLI tools to resolve conflicts, since that can make a rebase hang essentially hang while waiting for vim and printing a bunch of ANSI escape codes to stdout which makes the command output in ISL look like garbage, and can be hard to debug as a user.

To avoid this, we need to consider if the configured tool has set `.gui` True in its `merge-tools` config. This seems to match up with the exact criteria we need (although I think the original intention is sort of reversed: whether the system you're running on HAS a GUI to run this tool with...). Anyway, this lets us avoid ever trying to use `vimdiff` as the merge tool from ISL.

**Is this a reasonable approach to automatically running the merge tool? Would love feedback about this.**

Another aproach might be to have `sl` sniff if it's called from ISL, and make the merge tool selection ignore non-GUI tools there. There's already a `util.gui()`, but it's not detecting whether sl was run by a GUI, but rather if the system has a GUI at all. I kind of think this is a little messy either way—so it seems ok to put this in ISL itself where we have more control.

Reviewed By: quark-zju

Differential Revision: D57172364

fbshipit-source-id: 4c36d37d021361f276b3e8b16dca675a8555223b