Consolidate merge tool detection
Summary:
In the last diff, I had operations detect the merge tool being used, notably by checking if it sets .gui = True. We should use this same logic to check the merge tool when showing in the UI if a merge tool has been configured.

This means the UI no longer just uses `ui.merge`, but needs to send a custom message to find the merge tool.

This check should be memoizable, which we can do in another diff.

Reviewed By: quark-zju

Differential Revision: D57174605

fbshipit-source-id: 1462d9997ddeca8a9e46fe27e9e93ba04a5f7a2f