Cache merge tool
Summary:
We're finding the configured merge tool with an extra lookup for the `merge-tools.$tool.gui`. No reason to do this every time you run an operation, we can simply cache it in the RepositoryContext.

This state is already something that needs a restart to take effect when changed, since we cache `ui.merge`, so I think it's good to cache this too.

Reviewed By: quark-zju

Differential Revision: D57175640

fbshipit-source-id: e1b5e5c7d378561d91cbc70e1b726f47885624a6