commit_graph: return boundary changesets in ancestors_within_distance
Summary: Currently there's an `ancestors_within_distance` method that returns a stream of all ancestors of a set of changesets that are within a given distance `max_distance`. This diff renames this method into `ancestors_within_distance_stream` and makes it return for each changeset the minimum distance it takes to reach it alongside its changeset id. Another non-stream version is added `ancestors_within_distance` that returns the ancestors as well as all boundary changesets, which are changesets for which the minimum distance to reach them is exactly the given `max_distance`.

Reviewed By: mitrandir77

Differential Revision: D57280689

fbshipit-source-id: 3ed53049ab3b24e44482160417222f70449b93e8