Modify ancestors_within_distance to remove overlap between ancestors and boundary commits
Summary:
D57280689 added the `ancestors_within_distance` endpoint that provides two pieces of information. `Ancestors` which represent the ancestors of `heads` whose distance is less than `depth` and `Boundaries` which represents commits that are at a minimum distance `depth` from the `heads`. The current implementation defined `ancestors` to be inclusive of non-boundary and boundary ancestors. W.r.t the git usage, it would make operations quite inefficient if the `ancestors` property also included the `boundary` commits.

This diff updates the `ancestors_within_distance` method to ensure that we return boundary and non-boundary ancestors separately.

Reviewed By: YousefSalama

Differential Revision: D57333286

fbshipit-source-id: b6229f8bf912b635f37b3751680b31687fface92