Auto-resolve conflicts when saving a file
Summary: In VS Code, when saving a file (that's handled by an sl repository) during a merge conflict, where that file is part of the conflict info, but no longer has conflict markers, we can automatically run `sl resolve --mark file.txt` so you don't have to do it yourself. This should be a pretty safe thing to do, and it really speeds up the resolving process if you're resolving the files in the vscode editor.

Reviewed By: zzl0

Differential Revision: D57301327

fbshipit-source-id: 86267cbb4a0dffe31cd600eb86b252bb1f27f6db