Enable scuba logging by default in forward syncer
Summary:
## This stack
See D57204338.

## This diff
Basically reverts D53662810, enabling log by default again.

The original (pretty old) behaviour was to not log by default. I had to enable it by default to migrate forward syncer to new clap and
Mononoke app.
Later I disabled again, but **the problem** is that it happened a few times that we forgot to pass the scuba logging argument when running the forward syncer for merges (even prod merges).

It doesn't log a lot, so I figured it's best to enable by default so we don't risk losing data.

Reviewed By: RajivTS

Differential Revision: D57370495

fbshipit-source-id: f253e2dcc49f5231fb6a8c4d313a660c455d52bc