Synchronously derive all types during initial import
Summary:
## Context
As part of the merge process, we need to do the initial import of commits into the large repo.
We used to then manually backfill all derived data types for the imported commits.

In D56706360 we started automatically deriving fsnodes. See the summary and comments in this diff for why.

In D56877240 I made a change to automatically derive all the other data types asynchronously.
## Problem
The second change doesn't really work well in practice.

`derive_bulk` has the requirement that all the parents must have their data derived before deriving for a changeset.
Because we were calling it asynchronously, there were no guarantees that we would start and finish derivation of a changeset before calling it for its children.

## Fix
The time to backfill derived data will be spent anyway, so let's just do it synchronously during the initial import.
There are 2 clear benefits here:

1. This will save engineering time of manually running newadmin commands to backfill it later.
2. If there's are any unexpected issues during derivation of any of the types, this will stop the import and let us debug the issue to see if we should fix something or can proceed.

Reviewed By: mitrandir77

Differential Revision: D57371061

fbshipit-source-id: 1919f2c07996f813014d9333e51208eb54e855aa