change CFM rule to only move internal docs into the wiki
Summary:
Out of all of the docs that we have in the OSS wiki, we probably wouldn't like to put all of them into the wiki since many of those refer to OSS workflows.
We are particularly interested into the internals one, since we already have some old sections in the wiki that used to be synchronized through Sphinx at:
https://www.internalfb.com/intern/wiki/Sapling-Internal/dev/DeveloperGuide/

However, Sphinx is dead and we don't have a way to update these wikis, so we need a way to update them somehow. The plan is to move some of those remaining into the website and to use CFM instead to sync that.
We'll probably need to nuke the page above before landing this diff, so heads up for that.

Reviewed By: muirdm

Differential Revision: D56914755

fbshipit-source-id: 1bb6037feb1418615accbb232df8eb93dbea734f