Wait for running commands before running export stack when loading split UI
Summary:
Previously, if you clicked "split" or "edit stack", it would immediately run `debugexportstack`, without waiting for queueing to happen. This can lead to subtle issues, like the split UI not reflecting your last operation. It can also try to use optimistic commit hashes (though that might not be fixed by this diff....)

To improve this, we want to wait when loading the split/edit stack UI until all your queued commands have finished running.

This can technically introduce unnecessary waiting, for example if you're running `pull`, which shouldn't affect the current stack's commit contents.

I still think it's worth doing this, for consistency, and since it's somewhat hard to know which operations we do need to wait for. Also, by having the UI let you start the split, but wait to show the value, it behaves similarly to queueing, which feels very consistent to me. (contrast this with disabling the split button, which would make you "manually" wait for the operations to finish, then click again)

Reviewed By: quark-zju

Differential Revision: D57395177

fbshipit-source-id: 59f5d791b40f214a1035042f03096ef563893554